lesson2/requirements.txt
python-dotenv
requests==2.25.1
